add variable.go and rename uint32.go to element.go
