+++ utils/fingerprint.go
@@ -13,7 +13,6 @@ import (
 	"github.com/consensys/gnark/std/algebra/emulated/sw_bn254"
 	"github.com/consensys/gnark/std/commitments/kzg"
 	"github.com/consensys/gnark/std/math/emulated"
-	"github.com/consensys/gnark/std/math/uints"
 	"github.com/consensys/gnark/std/recursion/plonk"
 )
 
@@ -125,7 +124,7 @@ func (fp FingerPrint) IsEqual(api frontend.API, other FingerPrint) frontend.Vari
 func FpValueOf(api frontend.API, v frontend.Variable, bitsPerVar int) (FingerPrint, error) {
 	bits := api.ToBinary(v)
 
-	vals := bitsToVars(api, bits, bitsPerVar)
+	vals := BitsToVars(api, bits, bitsPerVar)
 	return FingerPrint{
 		Vals:       vals,
 		BitsPerVar: bitsPerVar,
@@ -174,38 +173,6 @@ func AssertFpInSet(api frontend.API, fp frontend.Variable, fpSet []FingerPrintBy
 	api.AssertIsEqual(sum, 1)
 }
 
-func ValsToU8s(api frontend.API, vals []frontend.Variable, bitsPerVar int) ([]uints.U8, error) {
-	uapi, err := uints.New[uints.U32](api)
-	if err != nil {
-		return nil, err
-	}
-
-	bytesPerVar := bitsPerVar / 8
-	ret := make([]uints.U8, bytesPerVar*len(vals))
-	for i := 0; i < len(vals); i++ {
-		bytes := uapi.ByteArrayValueOf(vals[i], bytesPerVar)
-		begin := i * bytesPerVar
-		end := begin + bytesPerVar
-		copy(ret[begin:end], bytes)
-	}
-
-	return ret, nil
-}
-
-func ValsFromBytes(data []byte, bitsPerVar int) []frontend.Variable {
-	bytesPerVar := (bitsPerVar + 7) / 8
-	ret := make([]frontend.Variable, 0)
-	for i := 0; i < len(data); i += bytesPerVar {
-		tmp := make([]byte, bytesPerVar)
-		copy(tmp, data[i:i+bytesPerVar])
-		slices.Reverse[[]byte](tmp)
-		ret = append(ret, tmp)
-	}
-
-	slices.Reverse[[]frontend.Variable](ret)
-	return ret
-}
-
 func AssertValsWithWitnessElements[FR emulated.FieldParams](
 	api frontend.API, vars []frontend.Variable, els []emulated.Element[FR], nbMaxBitsPerVar ...uint,
 ) {
@@ -249,16 +216,4 @@ func areVarsEquals(api frontend.API, a, b []frontend.Variable) frontend.Variable
 	}
 
 	return sum
-}
-
-func bitsToVars(api frontend.API, bits []frontend.Variable, bitsPerVar int) []frontend.Variable {
-
-	vals := make([]frontend.Variable, 0)
-	for i := 0; i < len(bits); i += bitsPerVar {
-		val := api.FromBinary(bits[i : i+bitsPerVar]...)
-		vals = append(vals, val)
-	}
-
-	slices.Reverse[[]frontend.Variable](vals)
-	return vals
 }
